pyproject.toml
[project]
name = "vittcoin"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111,<0.116",
  "uvicorn[standard]>=0.23,<0.33",
  "SQLAlchemy>=2.0,<2.1",
  "aiosqlite>=0.19,<0.21",
  "pydantic[email]>=2.7,<3.0",
  "pydantic-settings>=2.3,<3.0",
  "fastapi-cli>=0.0.5,<1.0",
  "email-validator>=2.3.0",
  "greenlet>=3.2.4",
  "psycopg2-binary>=2.9",
  "asyncpg>=0.29.0",
  "pyjwt[crypto]>=2.8.0",
  "google-auth>=2.29.0",
  "requests>=2.31.0",
  "httpx>=0.27.0",
  "alembic>=1.17.2",
]
